requirements.txt
mkdocs>=1.6,<2
mkdocs-material>=9.5,<10
mkdocs-awesome-pages-plugin>=2.9,<3
pyyaml>=6.0,<7
